Support 256-color and 24-bit RGB modes
Add support for specifying 256-color and 24-bit RGB color according to the
ISO and Xterm standards.  Colors specified as a single decimal number in
the range 0..255 will be treated as 256-color mode colors (with a leading
'on_' indicating a background rather than foreground color).  An RGB triplet
may be specified with three comma-separated decimal numbers, with a leading
'on_' working as it does with 256-color mode.  For example, 'on_128,192,0'
indicates a background color specified as a 24-bit RGB triplet.